Refactor modules into subpackages
