+++ snekbox/snekio/errors.py
@@ -1,0 +1,6 @@
+class ParsingError(ValueError):
+    """Raised when an incoming content cannot be parsed."""
+
+
+class IllegalPathError(ParsingError):
+    """Raised when a request file has an illegal path."""
