--- snekbox/utils/__init__.py
@@ -1,3 +1,0 @@
-from . import cgroup, logging, swap, timed
-
-__all__ = ("cgroup", "logging", "swap", "timed")
